pyproject.toml
[tool.poetry]
name = "gh-actions-exporter"
version = "0.1.0"
description = ""
authors = ["Release Engineering <release.engineering@scality.com>"]

[tool.poetry.scripts]
start = "gh_actions_exporter.main:start"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
prometheus-client = "^0.12.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
python-multipart = "^0.0.5"
PyYAML = "^6.0"
pygithub = "^1.58.1"
jinja2 = "^3.1.2"
pyjwt = "^2.6.0"
pytest-mock = "^3.10.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.1.6"
mkdocs-multirepo-plugin = "^0.6.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
flake8 = "^4.0.1"
requests = "^2.26.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
